requirements.txt
asgiref==3.8.1
Django==5.2.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
